expense_tracker/requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.1.5
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
sqlparse==0.5.3
tzdata==2025.1
whitenoise